+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 Dev
+# WW 05.12.2014
+* ssl certificat checker
+** plugin for ios and android
+** checks on dev/stage against *.cameo.io certificate
+** checks on prod against cameonet.de certificate
+** if fingerprint doesn't match the cmConnectionHandler shows a modal and blocked the app
+
 # BS 03.12.2014
 * update Event Handling CAM-1422 
 ** add "contact:update" in FE
@@ -6,7 +13,9 @@ Dev
 ** add "identity:new" in FE
 
 0.4.4.
-# BS 04.12.2014
+# BS 05.12.2014
+* fix CAM-1398, error will now shown, if key saving failed
+# BS 0..12.2014
 * fix CAM-1482, set timeout for browser notifications to 5s
 * fix CAM-1483, set z-index to choose element
 * fix CAM-1481, only one browser notification per message